+++ pyogrio/_io.pyx
@@ -402,7 +402,7 @@ cdef get_total_bounds(OGRLayerH ogr_layer, int force):
 
     except CPLE_BaseError:
         bounds = None
-    
+
     return bounds
 
 
@@ -595,8 +595,8 @@ cdef apply_where_filter(OGRLayerH ogr_layer, str where):
         raise ValueError(f"Invalid SQL query for layer '{OGR_L_GetName(ogr_layer)}': '{where}'")
 
 
-cdef apply_spatial_filter(OGRLayerH ogr_layer, bbox):
-    """Applies spatial filter to layer.
+cdef apply_bbox_filter(OGRLayerH ogr_layer, bbox):
+    """Applies bounding box spatial filter to layer.
 
     Parameters
     ----------
@@ -616,6 +616,28 @@ cdef apply_spatial_filter(OGRLayerH ogr_layer, bbox):
     OGR_L_SetSpatialFilterRect(ogr_layer, xmin, ymin, xmax, ymax)
 
 
+cdef apply_geometry_filter(OGRLayerH ogr_layer, wkb):
+    """Applies geometry spatial filter to layer.
+
+    Parameters
+    ----------
+    ogr_layer : pointer to open OGR layer
+    wkb: WKB encoding of geometry
+    """
+
+    cdef OGRGeometryH ogr_geometry = NULL
+    cdef unsigned char *wkb_buffer = wkb
+
+    err = OGR_G_CreateFromWkb(wkb_buffer, NULL, &ogr_geometry, len(wkb))
+    if err:
+        if ogr_geometry != NULL:
+            OGR_G_DestroyGeometry(ogr_geometry)
+        raise GeometryError("Could not create mask geometry") from None
+
+    OGR_L_SetSpatialFilter(ogr_layer, ogr_geometry)
+    OGR_G_DestroyGeometry(ogr_geometry)
+
+
 cdef validate_feature_range(OGRLayerH ogr_layer, int skip_features=0, int max_features=0):
     """Limit skip_features and max_features to bounds available for dataset.
 
@@ -632,13 +654,10 @@ cdef validate_feature_range(OGRLayerH ogr_layer, int skip_features=0, int max_fe
     num_features = max_features
 
     if feature_count == 0:
-        name = OGR_L_GetName(ogr_layer)
-        warnings.warn(f"Layer '{name}' does not have any features to read")
         return 0, 0
 
-    # validate skip_features, max_features
     if skip_features < 0 or skip_features >= feature_count:
-        raise ValueError(f"'skip_features' must be between 0 and {feature_count-1}")
+        skip_features = feature_count
 
     if max_features < 0:
         raise ValueError("'max_features' must be >= 0")
@@ -833,6 +852,10 @@ cdef get_features(
     ]
 
     field_data_view = [field_data[field_index][:] for field_index in range(n_fields)]
+
+    if num_features == 0:
+        return fid_data, geometries, field_data
+
     i = 0
     while True:
         try:
@@ -1043,6 +1066,7 @@ def ogr_read(
     int max_features=0,
     object where=None,
     tuple bbox=None,
+    object mask=None,
     object fids=None,
     str sql=None,
     str sql_dialect=None,
@@ -1065,10 +1089,10 @@ def ogr_read(
     path_c = path_b
 
     if fids is not None:
-        if where is not None or bbox is not None or sql is not None or skip_features or max_features:
+        if where is not None or bbox is not None or mask is not None or sql is not None or skip_features or max_features:
             raise ValueError(
-                "cannot set both 'fids' and any of 'where', 'bbox', 'sql', "
-                "'skip_features' or 'max_features'"
+                "cannot set both 'fids' and any of 'where', 'bbox', 'mask', "
+                "'sql', 'skip_features' or 'max_features'"
             )
         fids = np.asarray(fids, dtype=np.intc)
 
@@ -1080,6 +1104,9 @@ def ogr_read(
             "at least one of read_geometry or return_fids must be True or columns must "
             "be None or non-empty"
         )
+
+    if bbox and mask:
+        raise ValueError("cannot set both 'bbox' and 'mask'")
 
     try:
         dataset_options = dict_to_options(dataset_kwargs)
@@ -1152,7 +1179,10 @@ def ogr_read(
 
             # Apply the spatial filter
             if bbox is not None:
-                apply_spatial_filter(ogr_layer, bbox)
+                apply_bbox_filter(ogr_layer, bbox)
+
+            elif mask is not None:
+                apply_geometry_filter(ogr_layer, mask)
 
             # Limit feature range to available range
             skip_features, num_features = validate_feature_range(
@@ -1211,6 +1241,7 @@ def ogr_open_arrow(
     int max_features=0,
     object where=None,
     tuple bbox=None,
+    object mask=None,
     object fids=None,
     str sql=None,
     str sql_dialect=None,
@@ -1229,6 +1260,9 @@ def ogr_open_arrow(
     cdef ArrowArrayStream stream
     cdef ArrowSchema schema
 
+    IF CTE_GDAL_VERSION < (3, 6, 0):
+        raise RuntimeError("Need GDAL>=3.6 for Arrow support")
+
     path_b = path.encode('utf-8')
     path_c = path_b
 
@@ -1238,9 +1272,15 @@ def ogr_open_arrow(
     if fids is not None:
         raise ValueError("reading by FID is not supported for Arrow")
 
-    if skip_features or max_features:
+    IF CTE_GDAL_VERSION < (3, 8, 0):
+        if skip_features:
+            raise ValueError(
+                "specifying 'skip_features' is not supported for Arrow for GDAL<3.8.0"
+            )
+
+    if max_features:
         raise ValueError(
-            "specifying 'skip_features' or 'max_features' is not supported for Arrow"
+            "specifying 'max_features' is not supported for Arrow"
         )
 
     if sql is not None and layer is not None:
@@ -1251,6 +1291,9 @@ def ogr_open_arrow(
             "at least one of read_geometry or return_fids must be True or columns must "
             "be None or non-empty"
         )
+
+    if bbox and mask:
+        raise ValueError("cannot set both 'bbox' and 'mask'")
 
     reader = None
     try:
@@ -1300,7 +1343,10 @@ def ogr_open_arrow(
 
         # Apply the spatial filter
         if bbox is not None:
-            apply_spatial_filter(ogr_layer, bbox)
+            apply_bbox_filter(ogr_layer, bbox)
+
+        elif mask is not None:
+            apply_geometry_filter(ogr_layer, mask)
 
         # Limit to specified columns
         if ignored_fields:
@@ -1324,13 +1370,15 @@ def ogr_open_arrow(
         # make sure layer is read from beginning
         OGR_L_ResetReading(ogr_layer)
 
-        IF CTE_GDAL_VERSION < (3, 6, 0):
-            raise RuntimeError("Need GDAL>=3.6 for Arrow support")
-
         if not OGR_L_GetArrowStream(ogr_layer, &stream, options):
             raise RuntimeError("Failed to open ArrowArrayStream from Layer")
 
         stream_ptr = <uintptr_t> &stream
+
+        if skip_features:
+            # only supported for GDAL >= 3.8.0; have to do this after getting
+            # the Arrow stream
+            OGR_L_SetNextByIndex(ogr_layer, skip_features)
 
         # stream has to be consumed before the Dataset is closed
         import pyarrow as pa
@@ -1378,7 +1426,8 @@ def ogr_read_bounds(
     int skip_features=0,
     int max_features=0,
     object where=None,
-    tuple bbox=None):
+    tuple bbox=None,
+    object mask=None):
 
     cdef int err = 0
     cdef const char *path_c = NULL
@@ -1387,6 +1436,9 @@ def ogr_read_bounds(
     cdef OGRLayerH ogr_layer = NULL
     cdef int feature_count = 0
     cdef double xmin, ymin, xmax, ymax
+
+    if bbox and mask:
+        raise ValueError("cannot set both 'bbox' and 'mask'")
 
     path_b = path.encode('utf-8')
     path_c = path_b
@@ -1404,7 +1456,10 @@ def ogr_read_bounds(
 
     # Apply the spatial filter
     if bbox is not None:
-        apply_spatial_filter(ogr_layer, bbox)
+        apply_bbox_filter(ogr_layer, bbox)
+
+    elif mask is not None:
+        apply_geometry_filter(ogr_layer, mask)
 
     # Limit feature range to available range
     skip_features, num_features = validate_feature_range(ogr_layer, skip_features, max_features)