+++ pyogrio/geopandas.py
@@ -1,3 +1,5 @@
+import os
+
 import numpy as np
 
 from pyogrio._compat import HAS_GEOPANDAS, PANDAS_GE_20
@@ -74,11 +76,13 @@ def read_dataframe(
     max_features=None,
     where=None,
     bbox=None,
+    mask=None,
     fids=None,
     sql=None,
     sql_dialect=None,
     fid_as_index=False,
-    use_arrow=False,
+    use_arrow=None,
+    arrow_to_pandas_kwargs=None,
     **kwargs,
 ):
     """Read from an OGR data source to a GeoPandas GeoDataFrame or Pandas DataFrame.
@@ -111,14 +115,14 @@ def read_dataframe(
         If the geometry has Z values, setting this to True will cause those to
         be ignored and 2D geometries to be returned
     skip_features : int, optional (default: 0)
-        Number of features to skip from the beginning of the file before returning
-        features.  Must be less than the total number of features in the file.
-        Using this parameter may incur significant overhead if the driver does
-        not support the capability to randomly seek to a specific feature,
-        because it will need to iterate over all prior features.
+        Number of features to skip from the beginning of the file before
+        returning features.  If greater than available number of features, an
+        empty DataFrame will be returned.  Using this parameter may incur
+        significant overhead if the driver does not support the capability to
+        randomly seek to a specific feature, because it will need to iterate
+        over all prior features.
     max_features : int, optional (default: None)
-        Number of features to read from the file.  Must be less than the total
-        number of features in the file minus skip_features (if used).
+        Number of features to read from the file.
     where : str, optional (default: None)
         Where clause to filter features in layer by attribute values. If the data source
         natively supports SQL, its specific SQL dialect should be used (eg. SQLite and
@@ -132,21 +136,30 @@ def read_dataframe(
         and used by GDAL, only geometries that intersect this bbox will be
         returned; if GEOS is not available or not used by GDAL, all geometries
         with bounding boxes that intersect this bbox will be returned.
+        Cannot be combined with ``mask`` keyword.
+    mask : Shapely geometry, optional (default: None)
+        If present, will be used to filter records whose geometry intersects
+        this geometry.  This must be in the same CRS as the dataset.  If GEOS is
+        present and used by GDAL, only geometries that intersect this geometry
+        will be returned; if GEOS is not available or not used by GDAL, all
+        geometries with bounding boxes that intersect the bounding box of this
+        geometry will be returned.  Requires Shapely >= 2.0.
+        Cannot be combined with ``bbox`` keyword.
     fids : array-like, optional (default: None)
         Array of integer feature id (FID) values to select. Cannot be combined
-        with other keywords to select a subset (``skip_features``, ``max_features``,
-        ``where``, ``bbox`` or ``sql``). Note that the starting index is driver and file
-        specific (e.g. typically 0 for Shapefile and 1 for GeoPackage, but can
-        still depend on the specific file). The performance of reading a large
-        number of features usings FIDs is also driver specific.
+        with other keywords to select a subset (``skip_features``,
+        ``max_features``, ``where``, ``bbox``, ``mask``, or ``sql``). Note that
+        the starting index is driver and file specific (e.g. typically 0 for
+        Shapefile and 1 for GeoPackage, but can still depend on the specific
+        file). The performance of reading a large number of features usings FIDs
+        is also driver specific.
     sql : str, optional (default: None)
-        The SQL statement to execute. Look at the sql_dialect parameter for
-        more information on the syntax to use for the query. When combined
-        with other keywords like ``columns``, ``skip_features``,
-        ``max_features``, ``where`` or ``bbox``, those are applied after the
-        SQL query. Be aware that this can have an impact on performance,
-        (e.g. filtering with the ``bbox`` keyword may not use
-        spatial indexes).
+        The SQL statement to execute. Look at the sql_dialect parameter for more
+        information on the syntax to use for the query. When combined with other
+        keywords like ``columns``, ``skip_features``, ``max_features``,
+        ``where``, ``bbox``, or ``mask``, those are applied after the SQL query.
+        Be aware that this can have an impact on performance, (e.g. filtering
+        with the ``bbox`` or ``mask`` keywords may not use spatial indexes).
         Cannot be combined with the ``layer`` or ``fids`` keywords.
     sql_dialect : str, optional (default: None)
         The SQL dialect the SQL statement is written in. Possible values:
@@ -165,10 +178,15 @@ def read_dataframe(
     fid_as_index : bool, optional (default: False)
         If True, will use the FIDs of the features that were read as the
         index of the GeoDataFrame.  May start at 0 or 1 depending on the driver.
-    use_arrow : bool, default False
+    use_arrow : bool, optional (default: False)
         Whether to use Arrow as the transfer mechanism of the read data
         from GDAL to Python (requires GDAL >= 3.6 and `pyarrow` to be
         installed). When enabled, this provides a further speed-up.
+        Defaults to False, but this default can also be globally overridden
+        by setting the ``PYOGRIO_USE_ARROW=1`` environment variable.
+    arrow_to_pandas_kwargs : dict, optional (default: None)
+        When `use_arrow` is True, these kwargs will be passed to the `to_pandas`_
+        call for the arrow to pandas conversion.
     **kwargs
         Additional driver-specific dataset open options passed to OGR.  Invalid
         options will trigger a warning.
@@ -193,7 +211,11 @@ def read_dataframe(
 
         https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html
 
-    """
+    .. _to_pandas:
+
+        https://arrow.apache.org/docs/python/generated/pyarrow.Table.html#pyarrow.Table.to_pandas
+
+    """  # noqa: E501
     if not HAS_GEOPANDAS:
         raise ImportError("geopandas is required to use pyogrio.read_dataframe()")
 
@@ -202,6 +224,9 @@ def read_dataframe(
     from geopandas.array import from_wkb
 
     path_or_buffer = _stringify_path(path_or_buffer)
+
+    if use_arrow is None:
+        use_arrow = bool(int(os.environ.get("PYOGRIO_USE_ARROW", "0")))
 
     read_func = read_arrow if use_arrow else read
     if not use_arrow:
@@ -220,6 +245,7 @@ def read_dataframe(
         max_features=max_features,
         where=where,
         bbox=bbox,
+        mask=mask,
         fids=fids,
         sql=sql,
         sql_dialect=sql_dialect,
@@ -229,7 +255,15 @@ def read_dataframe(
 
     if use_arrow:
         meta, table = result
-        df = table.to_pandas()
+
+        # split_blocks and self_destruct decrease memory usage, but have as side effect
+        # that accessing table afterwards causes crash, so del table to avoid.
+        kwargs = {"self_destruct": True}
+        if arrow_to_pandas_kwargs is not None:
+            kwargs.update(arrow_to_pandas_kwargs)
+        df = table.to_pandas(**kwargs)
+        del table
+
         if fid_as_index:
             df = df.set_index(meta["fid_column"])
             df.index.names = ["fid"]